+++ R/reverselog_trans.R
@@ -23,7 +23,9 @@ reverselog_trans <- function(base = 10) {
     trans <- function(x) -log(x, base)
     inv <- function(x) base^(-x)
 
-    scales::trans_new(paste0("reverselog-", format(base)), trans, inv,
-              scales::log_breaks(base = base),
-              domain = c(1e-100, Inf))
+    scales::trans_new(name = paste0("reverselog-", format(base)),
+                      transform = trans,
+                      inverse = inv,
+                      breaks = scales::log_breaks(n = 6, base = base),
+                      domain = c(1e-100, Inf))
 }