+++ R/scale_longitude.R
@@ -74,7 +74,7 @@ scale_y_latitude <- function(name = "", ticks = 30,
 
 #' @rdname scale_longitude
 #' @export
-scale_x_level <- function(name = "", expand = c(0, 0), trans = "reverselog", ...) {
+scale_x_level <- function(name = "", expand = c(0, 0), trans = "sa_height", ...) {
     ggplot2::scale_x_continuous(name = name, expand = expand,
                        trans = trans,
                        ...)
@@ -82,7 +82,7 @@ scale_x_level <- function(name = "", expand = c(0, 0), trans = "reverselog", ...
 
 #' @rdname scale_longitude
 #' @export
-scale_y_level <- function(name = "", expand = c(0, 0), trans = "reverselog", ...) {
+scale_y_level <- function(name = "", expand = c(0, 0), trans = "sa_height", ...) {
     ggplot2::scale_y_continuous(name = name, expand = expand,
                        trans = trans,
                        ...)